feat: move grand search plugin to server.
move grand search plugin to server from desktop.

Log: 
Task: https://pms.uniontech.com/task-view-256381.html
