+++ src/plugins/desktop/CMakeLists.txt
@@ -10,5 +10,4 @@ add_subdirectory(core/ddplugin-canvas)
 add_subdirectory(core/ddplugin-dbusregister)
 add_subdirectory(ddplugin-background)
 add_subdirectory(ddplugin-wallpapersetting)
-add_subdirectory(ddplugin-grandsearchdaemon)
 add_subdirectory(ddplugin-organizer)
