--- src/plugins/desktop/ddplugin-grandsearchdaemon/CMakeLists.txt
@@ -1,38 +1,0 @@
-cmake_minimum_required(VERSION 3.10)
-
-project(ddplugin-grandsearchdaemon)
-
-set(CMAKE_INCLUDE_CURRENT_DIR ON)
-
-
-FILE(GLOB_RECURSE SRC_FILES
-    "${CMAKE_CURRENT_SOURCE_DIR}/*.h"
-    "${CMAKE_CURRENT_SOURCE_DIR}/*.cpp"
-    "${CMAKE_CURRENT_SOURCE_DIR}/*.json"
-)
-
-add_library(${PROJECT_NAME}
-    SHARED
-    ${SRC_FILES}
-)
-
-add_definitions(-DGRANDSEARCHDAEMON_LIB_DIR="${CMAKE_INSTALL_FULL_LIBDIR}/dde-grand-search-daemon")
-
-set_target_properties(${PROJECT_NAME} PROPERTIES LIBRARY_OUTPUT_DIRECTORY ../../)
-
-find_package(Dtk COMPONENTS Widget REQUIRED)
-find_package(PkgConfig REQUIRED)
-pkg_check_modules(DFrameworkDBus REQUIRED dframeworkdbus )
-
-
-target_link_libraries(${PROJECT_NAME}
-    DFM::framework
-)
-
-#install library file
-install(TARGETS
-    ${PROJECT_NAME}
-    LIBRARY
-    DESTINATION
-    ${DFM_PLUGIN_DESKTOP_EDGE_DIR}
-)
