--- src/plugins/desktop/ddplugin-grandsearchdaemon/ddplugin_grandsearchdaemon_global.h
@@ -1,14 +1,0 @@
-// SPDX-FileCopyrightText: 2022 - 2023 UnionTech Software Technology Co., Ltd.
-//
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-#ifndef DDPLUGIN_GRANDSEARCHDAEMON_GLOBAL_H
-#define DDPLUGIN_GRANDSEARCHDAEMON_GLOBAL_H
-
-#define DDP_GRANDSEARCHDAEMON_NAMESPACE ddplugin_grandsearchdaemon
-
-#define DDP_GRANDSEARCHDAEMON_BEGIN_NAMESPACE namespace DDP_GRANDSEARCHDAEMON_NAMESPACE{
-#define DDP_GRANDSEARCHDAEMON_END_NAMESPACE }
-#define DDP_GRANDSEARCHDAEMON_USE_NAMESPACE using namespace DDP_GRANDSEARCHDAEMON_NAMESPACE;
-
-#endif // DDPLUGIN_GRANDSEARCHDAEMON_GLOBAL_H
