--- src/plugins/desktop/ddplugin-grandsearchdaemon/grandsearchdaemon.json
@@ -1,24 +1,0 @@
-{
-    "Name" : "ddplugin-grandsearchdaemon",
-    "Version" : "1.0.0",
-    "CompatVersion" : "1.0.0",
-    "Vendor" : "The Uniontech Software Technology Co., Ltd.",
-    "Copyright" : "Copyright (C) 2022 Uniontech Software Technology Co., Ltd.",
-    "License" : [
-        "Licensees holding valid Qt Commercial licenses \
-            may use this plugin in accordance with the Qt Commercial License Agreement \
-            provided with the Software or, alternatively, in accordance with the terms contained \
-            in a written agreement between you and The Qt Company.",
-        "Alternatively, this plugin may be used under the terms \
-            of the GNU General Public License version 3 as published by \
-            the Free Software Foundation with exceptions as appearing \
-            in the file LICENSE.GPL3-EXCEPT included in the packaging of this plugin. \
-            Please review the following information to ensure the GNU General Public License requirements \
-            will be met: https://www.gnu.org/licenses/gpl-3.0.html."
-    ],
-    "Category" : "",
-    "Description" : "The common plugin for the filemanager and desktop.",
-    "UrlLink" : "https://www.uniontech.com",
-    "Depends" : [
-    ]
-}
